Move resource loader to shared library
